add topkgating to dlblas
