+++ python/dlBLAS/benchmarks/topk_gating.py
@@ -4,8 +4,20 @@ import torch
 import torch.nn.functional as F
 from torch import Tensor
 import triton
+import time
 
-from topk_gating import TopKGatingFunc, _capacity, gumbel_rsample
+from dlblas import get_op, get_list_op_names
+
+def _capacity(gates: Tensor, capacity_factor: Tensor, min_capacity: Tensor) -> Tensor:
+    # gates has shape of SE
+    num_tokens = gates.shape[0]
+    num_experts = gates.shape[1]
+    # to(torch.int64) works around a bug in torch.onnx.export:
+    # it should cast k to int64 when converting torch.topk but it doesn't.
+    capacity = torch.ceil((num_tokens / num_experts) * capacity_factor).to(torch.int64)
+    if capacity < min_capacity:
+        capacity = min_capacity.to(torch.int64)
+    return capacity
 
 def fused_topkgating(
     logits: Tensor,
@@ -15,6 +27,7 @@ def fused_topkgating(
 ) -> Tuple[Tensor, Tensor, Tensor, Tensor]:
     """Implements TopKGating on logits."""
     # everything is in fp32 in this function
+   
     gates = F.softmax(logits, dim=1)
     
     num_experts = int(gates.shape[1])
@@ -53,6 +66,7 @@ def fused_topkgating(
     # Store the capacity location for each token
     locations_s = torch.sum(locations * masks, dim=2)
    
+    
 
     # Normalize gate probabilities
     mask_float = masks.type_as(logits)
@@ -66,24 +80,46 @@ def fused_topkgating(
 
     # Calculate combine_weights and dispatch_mask
     gate_all = torch.einsum("ks,kse->kse", gate_s, mask_float)
-
-    # print(f"locations:{locations_s.shape}")
+    
+    # print(f"capacity:{capacity}")
+    
+    
+    # ---- test begin ----
+    # k, s, e, c= locations_s.shape[0], locations_s.shape[1], logits.shape[1], capacity
+    # combine_weights_test = torch.zeros((s, e, c), device=logits.device, dtype=logits.dtype)
+    # for idx_k in range(k):
+    #     for idx_s in range(s):
+    #         combine_weights_test[idx_s,:,locations_s[idx_k][idx_s]] += gate_all[idx_k, idx_s,:]
+    # --replace---
     locations_sc = F.one_hot(locations_s, num_classes=capacity).type_as(logits)
     combine_sec = torch.einsum("kse,ksc->ksec", gate_all, locations_sc)
     combine_weights = torch.sum(combine_sec, dim=0)
+    
+    # assert torch.allclose(combine_weights, combine_weights_test)
+    # --- test end ----
+    
+    # torch.cuda.synchronize(logits.device)
+    # t0 = time.time()
+
     dispatch_mask = combine_weights.bool()
-   
+
+    # torch.cuda.synchronize(logits.device)
+    # print(f"torch time:{(time.time() - t0) * 1000.0}")
+    
+    # return l_aux, masks, locations_s, exp_counts
     return l_aux, combine_weights, dispatch_mask, exp_counts
 
 
 
 def test():
-    device_ = torch.device('cuda:5')
+    device_ = torch.device('cuda:6')
     torch.cuda.set_device(device_)
-    # k, SeqLen, NumberExperts = 3, 4, 8
-    k, SeqLen, NumberExperts = 6, 1024, 64
+    k, SeqLen, NumberExperts = 8, 33, 64
+    # k, SeqLen, NumberExperts = 2, 4096, 64
     shape = (SeqLen, NumberExperts)
     logits_torch = torch.randn(shape, device=device_, requires_grad=True)
+    capacity_factor: float = 1.0
+    min_capacity: int = 2
     
     with torch.no_grad():
         logits_triton = logits_torch.clone()
@@ -91,11 +127,13 @@ def test():
     logits_triton.requires_grad = True
     
     model_torch = fused_topkgating
-    model_triton = TopKGatingFunc.apply
+
+    model_triton = get_op("topk_gating", (logits_triton, k, capacity_factor, min_capacity))
     
+    output1_torch, output2_torch, output3_torch, output4_torch = model_torch(logits_torch, k, capacity_factor, min_capacity)
+    output1_triton, output2_triton, output3_triton, output4_triton = model_triton(logits_triton, k, capacity_factor, min_capacity)
+
     
-    output1_torch, output2_torch, output3_torch, output4_torch = model_torch(logits_torch, k)
-    output1_triton, output2_triton, output3_triton, output4_triton = model_triton(logits_triton, k)
     
     assert output1_torch.shape == output1_triton.shape
     assert torch.allclose(output1_torch, output1_triton)
@@ -141,8 +179,8 @@ def test():
         ))
     @triton.testing.perf_report(configs)
     def bench_top2gating(SeqLen, Experts, provider, device=device_):
-        warmup = 100
-        rep = 100
+        warmup = 10
+        rep = 10
         shape = (SeqLen, Experts)
         logits = torch.randn(shape, device=device, requires_grad=True)
 
@@ -161,5 +199,8 @@ def test():
 
 
 
-test()
-print("sucessfully!")
+
+
+if __name__ == '__main__':
+    test()
+    print("sucessfully!")
